+++ README.md
@@ -1,8 +1,11 @@
-# resumable.ts
+# resumable-uploads
+
 A typescript supported version of the resumbalejs repo. The resumable.js repo is something that I use quite often. However it hasn't been updated or maintained since 2018. This project is to pick up where that left off. Fix bugs and other issues that were brought up and never fixed. Hopefully build a community of developers who can help maintain and upgrade this so that it can continue to be of use for many years to come
 
 ## Documentation
+
 A full break down of documentation can be found in [DOCUMENTATIION.md](./DOCUMENTATION.md)
 
 ## Contributing
+
 A [CONTRIBUTING.md]() file will be created soon with guidelines and requirements on how to contribute!
