+++ package.json
@@ -1,9 +1,9 @@
 {
-  "name": "resumablets",
+  "name": "resumable-uploads",
   "version": "0.1.0",
   "description": "A JavaScript library for providing multiple simultaneous, stable, fault-tolerant and resumable/restartable uploads via the HTML5 File API that supports typescript.",
-  "main": "resumablets.js",
-  "types": "./resumable.d.ts",
+  "main": "resumable-uploads.js",
+  "types": "./resumable-uploads.d.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "pre-commit": "lint-staged",
@@ -11,19 +11,19 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/augustcodes08/resumable.ts.git"
+    "url": "git+https://github.com/augustcodes08/resumable-uploads"
   },
   "keywords": [
     "html5",
     "file",
     "upload"
   ],
-  "author": "https://github.com/augustcodes08/resumable.ts/graphs/contributors",
+  "author": "https://github.com/augustcodes08/resumable-uploads/graphs/contributors",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/augustcodes08/resumable.ts/issues"
+    "url": "https://github.com/augustcodes08/resumable-uploads/issues"
   },
-  "homepage": "https://github.com/augustcodes08/resumable.ts#readme",
+  "homepage": "https://github.com/augustcodes08/resumable-uploads#readme",
   "devDependencies": {
     "@commitlint/cli": "^19.6.0",
     "@commitlint/config-conventional": "^19.6.0",
